requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
loguru
pillow
httpx
